+++ Employee-Service/src/main/java/com/ebf/employee/controllers/EmployeeController.java
@@ -1,41 +1,64 @@
 package com.ebf.employee.controllers;
 
+import com.ebf.employee.controllers.exceptions.EmployeeNotFoundException;
 import com.ebf.employee.entities.Company;
 import com.ebf.employee.entities.Employee;
 import com.ebf.employee.repos.CompanyRepository;
 import com.ebf.employee.repos.EmployeeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Optional;
 
 @RestController
+@CrossOrigin(origins = "http://localhost:4200")
 @RequestMapping("/employees")
 public class EmployeeController {
     @Autowired
     EmployeeRepository employeeRepository;
+
     @Autowired
     CompanyRepository companyRepository;
 
     @GetMapping
-    public Iterable<Employee> getEmployees() {
-        return employeeRepository.findAll();
+    public Page<Employee> getEmployees(Pageable pageable, @RequestParam(name = "companyId", required = false) Optional<Long> companyId) {
+        if (companyId.isPresent()) {
+            return employeeRepository.findAllByCompanyId(pageable, companyId.get());
+        }
+        return employeeRepository.findAll(pageable);
     }
 
-    @GetMapping("/{company_id}")
-    public Iterable<Employee> getEmployeesByCompany(@PathVariable long company_id){
-        return employeeRepository.findAllByCompanyId(company_id);
+    @GetMapping("/{id}")
+    public Employee getEmployeeById(@PathVariable long id) throws EmployeeNotFoundException {
+        Optional<Employee> employee = employeeRepository.findById(id);
+        if (employee.isPresent()) {
+            return employee.get();
+        }
+
+        throw new EmployeeNotFoundException("Employee by id("+id+") not found");
     }
 
-    @PostMapping("/add")
-    public Iterable<Employee> addCompany(@RequestBody Employee employee) {
-        employeeRepository.save(employee);
-        return employeeRepository.findAll();
+    @PutMapping
+    public Employee addEmployee(@RequestBody Employee employee) {
+        return employeeRepository.saveAndFlush(employee);
     }
 
-    @PostMapping("/delete/{id}")
-    public Iterable<Employee> deleteCompany(@PathVariable Long id) {
-        employeeRepository.deleteById(id);
-        return employeeRepository.findAll();
+    @DeleteMapping("/{id}")
+    public boolean deleteEmployee(@PathVariable Long id) {
+        if(employeeRepository.findById(id).isPresent()){
+            employeeRepository.deleteById(id);
+            return true;
+        }
+        return false;
+    }
+    @PostMapping
+    public Employee updateEmployee(@RequestBody Employee employee) throws EmployeeNotFoundException{
+        if(employeeRepository.findById(employee.getId()).isPresent()){
+            return employeeRepository.saveAndFlush(employee);
+        }
+
+        throw new EmployeeNotFoundException("Employee by id("+employee.getId()+") not found for update");
     }
 }
